Update the Norwegian tables
The purpose of this update is mainly to fix some problems caused by an update
in April 2015, that led to misbehavior when used with devices that don't
understand unicode bbraille. Some characters are moved to get higher
priority to output from liblouis.

In addition, I've added/fixed braille representation for some capital
accented letters that previously was wrong. These chars are added to the test file
no_harness.yaml as well.

Still there are one issue that I hope someone else can look into: the
begcomp/endcomp construct. I've added a test case, with status xfail, in the
file no_backward.yaml, to demonstrate the problem. The problem is not
introduced with this update.